requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.1.1
psycopg2-binary>=2.9.9
pdfplumber>=0.11.0
google-generativeai>=0.7.0
python-dotenv>=1.0.0
werkzeug>=3.0.0
gunicorn>=21.2.0
